Fix bug with image-compute.
image-compute was failing with some complex programs.  It *appears* that
the problem was due to weird interactions with the Racket garbage collector.
Solved the problem by making a copy of the region of interest, which is
probably safer, anyway.

Also made a small rename of variables during readthrough with Dr. Davis.